Change landing page name, apple crisp to apple crumble, fix gradient
This commit changes the name of the landing page from "Odin's Recipes" to "Sam's Recipes". It changes the recipe name "Apple Crisp" to "Apple Crumble" for the UK name. It also fixes a problem on the Tiramisu recipe page where the gradient background repeated near the bottom of the page.